+++ R/01-workflowsteprun-class.R
@@ -8,8 +8,10 @@
 #' @param step   A `workflowstep` object representing the step definition.
 #' @param args   A list of arguments that were passed to the step's command.
 #'  Contains actual values used during execution (e.g. results of previous steps).
-#' @param output The output produced by the step, if successful.
-#' @param error  An error object if the step failed, otherwise `NULL`.
+#' @param output List of outputs produced by the step if it executed
+#'               successfully, otherwise list(NULL).
+#' @param error  List of error conditions if the step encountered errors,
+#'               otherwise list(NULL).
 #' @param run_id A unique identifier for the workflow run this step belongs to.
 #' @param ...    Additional metadata to store with the step run.
 #' @return A `workflowsteprun` object.